+++ src/cli/DRagLint.CLI.pas
@@ -3,7 +3,7 @@ unit DRagLint.CLI;
 interface
 
 const
-  VERSION = '0.21.0-alpha';
+  VERSION = '0.22.0-alpha';
 
 function Run: Integer;